test: Move test_middleware_wait to non-pytest file, since it raises "twisted.internet.error.ReactorAlreadyRunning"
