+++ tests/middlewares/test_delayed_request_middleware.py
@@ -1,10 +1,6 @@
-import time
-
 import pytest
 from scrapy import Request
-from scrapy.core.downloader import DownloaderMiddlewareManager
 from twisted.internet.defer import Deferred
-from twisted.trial.unittest import TestCase
 
 from kingfisher_scrapy.downloadermiddlewares import DelayedRequestMiddleware
 from tests import spider_with_crawler
@@ -21,34 +17,3 @@ def test_middleware_output(meta, expected):
     output = middleware.process_request(request, spider)
 
     assert isinstance(output, expected)
-
-
-def test_middleware_wait():
-
-    def download_func(spider, request):
-        return request
-
-    spider = spider_with_crawler(settings={
-        'DOWNLOADER_MIDDLEWARES': {
-            'kingfisher_scrapy.downloadermiddlewares.DelayedRequestMiddleware': 543,
-        },
-    })
-    request = Request('http://example.com', meta={'wait_time': 1})
-    # We send the request to all the downloader middlewares, including the delayed request middleware.
-    manager = DownloaderMiddlewareManager.from_crawler(spider.crawler)
-    downloaded = manager.download(download_func, request, spider)
-
-    assert isinstance(downloaded, Deferred)
-
-    start = time.time()
-
-    # https://github.com/scrapy/scrapy/blob/28262d4b241744aa7c090702db9a89411e3bbf9a/tests/test_downloadermiddleware.py#L36
-    results = []
-    downloaded.addBoth(results.append)
-    test = TestCase()
-    test._wait(downloaded)
-
-    end = time.time()
-
-    assert results == [request]
-    assert 1 <= end - start <= 1.01
